Bug 1158509 - Reading List item does not announce if it is read/unread with accessibility
We communicate that the item is read primarily by speaking it with lower
pitch. So a VoiceOver user immediately learns whether item is
read/unread by noting the pitch of the voice. As read/unread status is
not communicated by a word at the beginning of the accessibility label
and the label begins with the title of the webpage, the user will also
immediately learn the beginning words of the title of webpage. I.e., the
title of the webpage and read/unread status are learned simultaneously
and immediately by the VoiceOver user.

This allows for quick navigation through the items, as within first
fractions of a second VoiceOver user can identify both read/unread
status and the beginning of the title of the article.

"Read"/"Unread" is still announced after the title, in case the user is
using only braille display without speech, or the user has trouble
recognizing pitch changes, or the user is a Switch Control user (Switch
Control does not respect the UIAccessibilitySpeechAttribute* attributes
- <rdar://problem/20701917>).