Bug 1161327: Fix order of accessility elements in various places
VoiceOver traverses elements like when people read a page - top-down,
left-right. In some places, the frames of things in the UI are such that
this leads to very bad user experience with VoiceOver w.r.t. order in
which it goes through the UI.

E.g. the tabs button has frame whose top corner is above the location
label. This makes VoiceOver go to the tabs button before going to the
location label.

So make the order of elements explicit so that VoiceOver visits it
correctly and provides a good user experience.

We do this:
- in location view to fix the secure lock image, location label and
reader mode butotn order
- in url bar to fix the tabs button order
- in search suggestions to put proposed answers after the question :-)