Bug 1164231 - Part 1: Accessibility actions for location view
I.e. paste&go, paste, copy address.

As the actions can vary based on whether the pasteboard is empty or not,
let the BVC decide what actions are available and use delegation to get
the actions in LocationLabel.

Action abstraction introduced that can represent an "action" having a
"name" and "handler", that can then be used to construct both
UIAlertAction and UIAccessibilityCustomAction.

As VoiceOver is unable to execute a UIAccessibilityCustomAction subclass
holding an instance of Action (and nothing else holding it), we cannot
generate Actions "on-the-fly" and have to make them persistent (stored)
somewhere in the view or view controller hierarchy (in this case BVC).

We could in the future expand this abstraction to support actions that
know whether they are available or not, and make it a one-liner to
install an action or a list of actions on a view, which would handle at
the same time both accessibility and long-press (with UIAlertAction for
list of more than one action).