Bug 1214794 - Ensure whatever the number of tiles in the data set are, that the correct number of tiles are backfilled
We had a situation whereby if the number of synced tiles was > the number of previous displayed tiles but < the total allowable number of displayed tiles, everything crashed
Therefore we are working out the max number of tiles that should be displayed based on the data and ensuring that the correct number are deleted and added.
Also fixed situation whereby new tile count < previous tile count as existing test failed given new calculations.