Bug 1226406 - Part 1: add asynchronous connection and transaction operations to SwiftData.
Run code within a connection scope or transaction, but don't block waiting for it to return.

The current SwiftData withConnection and transaction calls, and their wrappers in BrowserDB, are the worst kind of async -- they block two threads at once!

We call dispatch_sync to hand a block over to the connection's queue.

That means we wait for the current queue contents to finish running, run the block, and then finally return from the caller.

And very often what we do after that is return a Deferred or invoke some UI callback.

In those cases, there's really no reason to use dispatch_sync; we can have the callback do the right thing, such as filling the Deferred. The calling code can roll right on and do what it likes, chaining off the Deferred if it needs the result.