Bug 1239028 – Ensure spotlight displays favicon on first time visit to page.
https://bugzilla.mozilla.org/show_bug.cgi?id=1239028

The primary cause of this bug is that once becomeCurrent() has been called, setting the thumbnail data doesn't make any effect.

The fix is to ensure that the we call becomeCurrent() only once, but we need to be able to detect when no favicons are available.

There is a corner case here, which is that if favicons are specified in the HTML, but they all fail to load, then the page won't be be available for Spotlight or handoff.