Bug 1267167 – Today widgets open URLs in correct version of Firefox. (brave/brave-ios#1876)
This introduces a second URL scheme which is supposed to be internal to the app. The firefox scheme still exists, but allows us to be more predictable with different versions of the Firefox.

The biggest change to be aware of is the schemes are validated against the schemes registered in Info.plist (which is a bit redundant), and that it gets rid of any mention of the firefox-x-callback scheme, which was unused.

The Today widget needs to do a bit more work in order to pick the correct scheme.

This leaves every version of the app registering to open firefox:// URLs, which means any version of the app will be compatible with the OpenInFirefox library.