Bug 1274290 - Using google search operations that have the format 'operation:' do nothing when searching
By only accepting URLs that have valid schemes (i.e. on the approved list of URI schemes), we can now make searching in the form “command:value” (used by Google, for instance). It would hypothetically need to be updated if this list changes, but as the list changes rarely (not considering the provisional schemes), this would not appear to be an issue.