Bug 1300766 - only run the reopening code on toolbar state trait collection changes if on an iPad, otherwise it will always re-open the menu on rotation always.
This is because the trait collection change code runs before the menu dismissal code, which means the menu is dismissed and re-opened for the new trait collection by the time the shouldCloseMenu check is performed, causing `shouldCloseMenu` to always return false, ensuring that the menu is not dismissed when it should be. By limiting the toolbar dismissal code to iPad only, this keeps the menu rendering in the correct place on iPads, but does not interfere in the menu dismissal code for iPhones and portrait->landscape menu rendering on iPad.