Prefer process.env.PATH over process.env.Path on Windows
Auditors: @bridiver

I'm not sure if this is because of minGW that I use but I have node.env.PATH and that needs to be used.

CC: @bsclifton

Please give it a try to make sure it still works for you. The best way
is just to do new clone and then `npm install` and `npm run init`. If it
doesn't fail right away it's good.  I think it'll be fine.