Bug 1384159: update SDWebImage to latest (4.1) (brave/brave-ios#3186)
* Bug 1384159: update SDWebImage to latest (4.1)
This allowed removing some old code, and removing
some forced unwrapping with '!'.

* Review updates and fix broken tests

* Fix test where comparing PNG's of UIImage
A byte-for-byte comparison of PNG data was failing, in what appeared to be the
PNG metadata block. If UIImages are constructed identically, only then does
byte-wise PNG data comparison work.

* DiskImageStoreTests needed updating (and was not running)
When updating this code, I saw it was not included in the project
and has not been running. Is included now.