Bug 1405043 — Trivial speed optimizations for XCUITests. (brave/brave-ios#3247) r=njpark
* Bug 1405043 — Add Skip What's New and Skip Intro to all tests.

This tidies up the launchArguments to just being used in the TestAppDelegate, instead of spread around AppConstants. Each of SKIP_WHATSNEW, SKIP_INTRO and CLEAR_PROFILE can be used independently of one another.

Additionally, this cleans up the `restart()` method to the home button instead of `terminate()` and use `activate()` instead of `launch()`. This does not affect test setUps, but does help with tests that require restarts.

* Bug 1405043 — Speed up the animations to 100 times as fast.