Bug 1169322: Use a temp table to speed up frecency query. (brave/brave-ios#3554)
* Bug 1169322: Use a temp table to speed up frecency query.
Multiple views are recalculated on every keystroke, but one can assume
the database is unchanging while in the autocomplete view state.
Thus, create a temporary table when the autocomplete state is entered to
reduce the number of queries.
Also, joining result sets with favicon tables is adding 15-20% more time
to the query, and the client code never uses this data in the result
table. The client view that shows the favicons (correctly) looks up the
favicon for a site as needed.

iPhone5 queries are 2.5-3 sec per keystroke, with this patch they are
0.4 sec consistently. There will be some overhead to create the temp
  table, however it is off main, and the overall perf is so much better
I don't notice any issues.

The next thing to do is to remove the favicon sql joins from the code.
I think it is a valuable cleanup task.

* OptimizedFrecency lacked a proper protocol

* Fix logic for delete/create and topsitesquery

* Rename to FrecentHistory

* Instatiate SearchLoader on-demand when search state entered.

* Fix tests

* Create FrecencyParams enum to encapsulate conditional arguments
for frecency query generation. Otherwise arguments can conflict such as
bookmarksLimit > 0 and top sites type query performed and produce bad SQL.

* Fix tests to match interface changes
Change top sites cache test to explicitly use top sites interfaces, not
url frecency interface.

* Address review comments

* Set searchLoader = nil in BVC when exit search mode