Add test data files
This is a collection of the test files we need to use.

The DAT files are a bit big, but I think it's a good idea to keep
similar DAT files to what we use in production, so I am OK with that
trade-off.

I also included HTTPSE and TrackingProtection DAT files for upcoming
tests here.