Change OSCrypt password when importing from Chrome
We want to reuse Chrome's code as much as possible when importing encrypted
data (e.g. saved passwords, cookies) from a Chrome profile. The encryption is
managed by OSCrypt, which has an unfortunate singleton design that makes it
difficult to modify its behavior at runtime. This commit uses Muon's approach
of using a command line switch as a proxy for a global variable to change the
behavior of OSCrypt in the importer helper process.