Copy test Chrome profile to tempdir for unittests
ChromeImporter::ImportPasswords modifies the "Login Data" database every time
it runs. This is a side effect of LoginDatabase::Init and is not easily fixed.

As a result, the ImportPasswords test modifes the test "Login Data" file, which
is under version control, every time it runs. Copying the test profile data to
a tempdir prevents the tests from inappropriately modifying files under version
control.