UI: Move bookmark button to primary navigation area
Adds a new brave-core Class: `BookmarkButton` (borrows from `StarView`)

Adds a new brave-core subclass for `ToolbarView`: `BraveToolbarView`, which creates the BookmarkButton and adds to Layout.
Adds a new brave-core subclass for `LocationBar` so that the original `StarView` is no longer displayed.

Patches `BookmarkBubbleView` to so it's left side is anchored to the BookmarkButton, since this has moved from the right side to the left side of the ToolbarView.
- set_arrow cannot be called in our own method without subclassing BookmarkBubbleView since the original `ShowBubble` method sets the arrow and Inits the bubble before returning. Once it is set it cannot be changed.

Note that there is a preference for whether the editing of bookmarks is enabled, though that seems to be automatically controller rather than user-controlled. For example, guest windows will have that setting disabled.