provisioning profile and entitlement support on mac
This was added upstream which ended up breaking the the signing scripts
and caused the osx build to not start.

We might be ok reverting the changes if necessary like in
https://github.com/brave/brave-core/commit/faed762d267041bf3b4a5b8b048a0c827509554a
but i'd prefer to not worry about what that might break down the line.

Upstream signing changes:
https://chromium.googlesource.com/chromium/src.git/+/4d80b33e7c64069cb69041c4957ea77ecf6e14e8

Original more descriptive but reverted commit:
https://chromium.googlesource.com/chromium/src.git/+/b69d8c48da838196d9655402dd739e62227762a7