Use a consistent bundle ID
This was causing tests to fail for the c69 minor upgrade. In particular the bundle ID on the first call was org.brave.Brave, and then a call was made to set the bundle ID which changed subsequent calls to org.brave.Brave.development.  This makes the bundle ID the same always depending on the channel.