Ensure ShieldsActionView is destroyed before its controller is destroyed.
Fixes a crash most apparent whilst closing windows on Linux where the ToolbarActionView destructor attempts to call into a member of the view controller.
Due to the ordering of the smart pointer definitions in brave_actions_container.h, the view controller was getting destroyed first. Instead of relying on this implicit behavior, we explicitly destroy in the destructor to ensure this is always the case.