Don't retrieve first run time on main thread
Retrieving the time is a blocking operation and will trigger an
assertion on the main thread. Instead, post a task to retrieve the
first run time and when it completes post tasks back to the main
thread to perform associated bookkeeping (these need to run on the
main thread as they interact with the preference service).