Add shadows to tabs
Adds tab painting function for shadows.

In order to paint the tab's (top) shadow above the tab's (previous) bounds rectangle, we must extend the tab (and tabstrip) higher, to the top of the window (usually 8px of space). This is because the Views framework does not allow painting outside a View's bounds.

A hacky alternative would be to pass the PaintInfo from the BrowserView through the TabStrip child to the Tab descendants, so that a canvas can be created that can span the larger window rectangle. Aside from that strategy not done anywhere else in Chromium, it could cause issues if the tabs end up on a different paint layer, as well as paint invalidation and caching issues that would need to be solved.

Close https://github.com/brave/brave-browser/issues/788