Update mini installer to parse referral code from command-line arg
The stub installer will parse its own filename and pass it to the mini
installer via the command line interface (it should be the last
argument on the command line). The mini installer will parse the
filename for a referral code and send it to the setup program as usual
if found. This will allow referrals to work via the updateable
installer.