Keep the menu items in private windows with Tor same as in others.
fix brave/brave-browser#1627

The result is a little silly -- because private windows with Tor are
implemented in a guest session, all new windows open in the same
session, whether with 'New window' or 'New incognito window' (soon to
be 'New private window') or 'New private window with Tor').

On the other hand, this is perhaps justifiable because we don't want
to make it easy to accidentally _stop_ using Tor.  You can do that
explicitly through the profile menu -- but not just by casually
hitting Ctrl-N.

On the third hand, control-click on links _greys out_ the 'Open link
in private window' item when you're in a private window, rather than
making it just...do what it says.  ('Open link in new window' also
opens in a new private window in that case.)  For consistency, maybe
we should do that instead, when we are on less of a deadline?