Promo code is never claimed if user closes browser too soon
We previously only looked for the promoCode file on first
run. However, since there's a short delay before we launch the task to
read the promoCode file, there's a window of opportunity for a user to
close the browser before we read the file. This puts the user in a
state where the first run has occurred but the promoCode file wasn't
read (and therefore won't ever be read).

We now use a Boolean preference to achieve the same effect, but since
it's not tied to first-run it will allow us to continue to look for
the promoCode file even in the above scenario. Once we've checked for
the promoCode file we flip the flag so that there's no long-term
cost.