Handle invalid grant notification ID
We recently changed the format of the grant notification ID from an
integer to a string. We've encountered a scenario where an old install
may erroneously set the ID to 0 (which was never a valid value). Once
you're in this state, it's impossible to dismiss the notification. To
fix this, we look for this case on startup and modify the notification
to have the correct ID.