Listen for referral for referral headers updates
We were previously only retrieving the referral headers on first
launch. Aside from preventing us from incorporating potential updates
to the headers, this also made it so that a browser restart was
required the first time a user initialized the referral (as the
referral headers would still be empty from the network delegate's
perspective).

Now we register for changes to that preference and pull them in on the
fly.