Dynamically pack and serve all webpack generated files
As a necessity for that, only build the required webpack entry-point for each target.

This creates a new GN target type "pack_web_resources", which takes the output from the transpile_web_ui target and builds in to a pack. We then take the `GritResourceMap` which is created and use it to ensure that each WebUI or Extension DataSource will serve all files that were generated.

The "transpile_web_ui" target now generates a GRD file to be consumed by this new target, which includes all the files in the webpack output dir.

We also re-pack to combine these generated assets along with their static counterparts (i.e. html file) to the single .pak file that other build targets are expecting.

Follow-ups to come afterwards which are enabled by this:
- Move most other static resources in to webpack-generated resources, especially images. This will reduce the amount of definitions that need to be kept in sync, ensure any optimizations are run, and ensure that we aren't including files that aren't used.
- Use Code Splitting to generate multiple JS files for each WebUI as appropriate.
- More consistent dir structure / targets for components/brave_xyz_ui.