Use forward declaration for GritResourceMap instead of copying generated definition
For now this imports a generated grit output file which is not actually a dependency, but will contain the accurate generated definition of GritResourceMap.

We should consider replacing this with the actual source GritResourceMap definition, if that is contained anywhere accessible in the codebase, or generate a small GRD and grit output in this GN project which can be more locally referenced.