Fix public asset path for webpack-generated assets in extensions
For external assets e.g. an .svg file, webpack emits the file in the fs at the path specified by target_gen_dir or target_gen_dir + extra_relative_path. webpack also outputs a relative URL to access these assets. By default this relative URL is '/' but if we do specify an extra relative path, then we need to specify a new public relative url to access those resources. In the case of extensions, this needs to be the portion of the extra_relative_path that is deeper than where the manifest.json lives. This is because chromium matches the extension resources with ones that have the same fs path prefix as the manifest.json path.