Add dep for tests
bbondy [4:58 PM]
trying to get tests working on 0.58.x and noticed this is failing for tests:
https://github.com/brave/brave-core/pull/924

I noticed this is in 0.57.x and 0.58.x but it needs a dependency for
which was added with this https://github.com/brave/brave-core/commit/f929cbc89589e1f7f149280c86361f99a644a3c3
now I'm just going to add the dep for now ,but I want to make sure this is not a problem
and check in case we needed the other PR in those branches too
or to revert 924 from them

petemill [5:04 PM]
hmm good spot. I don't think there is anything else that makes 924 depend on 774. We probably don't need 774 on 0.57 as it's not urgent. I don't know how urgent 924 is but it sounds like it could be if it was causing an infinite loop. So safest is probably to add the dep.

bbondy [5:06 PM]
great thx for confirming, added the dep