Prototype of User Tipping on twitter.com
After discussion, implemented in pure HTML to begin with in order to get a working version rapidly and unblock some backend implementation and UX review.

- Detects 'like' action via network requests
- Detects 'like' button position via DOM query
- Shows dialog via plain DIV injection
- Dialog content is shown via closed Shadow DOM

Uses JS and HTML to show the dialog inside the host content document.
Security precautions:
- All execution is done within an isolated V8 content script
- Does not expose any extension URLs to the host page (as would be the case with a src'ed iframe or chunked JS)
- Renders everything to a shadow DOM

One consideration and possible disadvantage of this method is whether the host site could fingerprint a brave user with the detection of this extra DOM element. However,
- This method renders a simple `<div style="position: absolute;" />` without providing the host any further identification vectors or access to DOM.