Issue 3334: Fix lint script for uplift requests
script/lint.py uses GetCommonAncestorWithUpstream() returns common ancestor
with the master branch. This function returns the top_commit if the changes
have been pushed to the branch. This change fixes thelint script by passing
a base branch as an argument to the lint script to get the list of files
changed by the commit.

auditors: @bridiver, @mihaiplesa, @bsclifton