Fixes clear browsing data on exit.
The functionality was called from override of
ChromeBrowserMainParts::PreShutdown, which was removed along with the
call to it from BrowserMainLoop::PreShutdown. This fix subclasses
BrowserMainLoop and calls BraveBrowserMainParts::PreShutdown from it.
It also patches BrowserMainParts to add virtual PreShutdown method
and BrowserMainRunnerImpl to instantiate BraveBrowserMainLoop instead
of BrowserMainLoop.

Chromium change:

commit 0f15e7a22545884fd259b873366a511686f4e5d9
Author: Alexei Filippov <alph@chromium.org>
Date:   Thu Feb 14 21:34:28 2019 +0000

    [heap profiler] Migrate UMA heap profiler off LegacyCallStackProfileBuilder

    ... and remove the latter as it was the last client.

    BUG=916303,913570