Allow IPFS Companion Beta to have chrome.sockets.*
The ID of the Stable channel was whitelisted in https://github.com/brave/brave-core/commit/10f16e6f5255bc2300cc8d30021894ec6e717442
but we want to allow the ID of Beta channel to have access to those APIs as well.

The purpose of opt-in Beta channel is to dogfood and iterate faster
without impacting regular user base.