Enable vaapi support
Partially fixes #1024 

By default browser will blacklist all drivers on linux for vaapi decoding. 

To enable it , need to pass `--ignore-gpu-blacklist` in order to get it working since libva has problems atm with intel GPUs. I am hoping that it will be fixed in the future versions. There are a lot of activities in upstream in this.
Only adding this support on 64 bit build at the moment.

 Test case:
Run brave with ignore-gpu-blacklist flag and see if the video is playing or not
Based on this upstream [commit](https://github.com/chromium/chromium/commit/31225b9c5f3f685d65f742dc129241c30c32157c)

Need to add libva 2.4 as a build and runtime dependency.