WebUI: Loading shim gets consistent brave navigation bar style on Settings, Bookmarks, Downloads, History and Extensions
Note that chrome://brave-resources/page_specific/X url-data-source is used for convenience. Technically these belong in //browser/resources and not //ui/resources, and served from, for example, chrome://bookmarks/X instead. However, the build setup neccessary to copy in to the chromium build output directory before the PAK and Vulcanize process is a lot of overhead for a couple lines of simple css. In the future we may wish to do this setup.