Fix BraveActionViewController.
ExtensionActionViewController now requires a valid instance of
ToolbarActionBar. To get around it patched ExtensionActionViewController
to not DCHECK for it and patched the class definition to allow for
overriding a number of methods that use the passed in instance of the
ToolbarActionBar. Overrode those in BraveActionViewController to not
call out into ToolbarActionBar.

Chromium change:

commit 4d71c399c2cfe34f45114e13ba83e293122ecf2c
Author: Peter Boström <pbos@chromium.org>
Date:   Wed Apr 10 00:21:04 2019 +0000

    Add extension pop-out to Extension menu

    This changes ExtensionActionViewController to be aware of a main
    ToolbarActionsBar, which it asks directly in
    ::GetPreferredPopupViewController() instead of delegating into the
    owner's ::GetMainControllerForAction(), which is now gone.

    Separating the main bar from the owning bar allows ExtensionMenuView to
    make use of it even though it's not a ToolbarActionsBar.

    Bug: chromium:943702