Add urllib3 retry logic for upload POST action
This should initiate 5 retries when we receive
502 Bad Gateway errors uploading to GitHub.

urllib3 is only installed on the sf-mac-pro build
worker, and we use sys.path.append to add the path
to the package because it doesn't find it by default.
We will use pip to install urllib3 on the linux
and windows AWS workers in the build job itself.

Fixes https://github.com/brave/devops/issues/943