Add persistent state state for Shields close https://github.com/brave/brave-browser/issues/4953
Shields state needed across windows/tabs is not stored in localStorage API.
This follows the webUI convention for persistent state.

Goal is to provide a way to handle upcoming changes such as onboarding screen
and simple/advanced view setting.

Other state such as stats, resources blocked, and blocking preferences are still
handled by the back-end and are left untouched.