Fixes seg fault crash of BraveViewsDelegateViewLinuxBrowserTest.GetDefaultWindowIconTest
As @simonhong noted in his comments: "In test, our BraveViewsDelegateLinux isn't set
because ChromeViewsDelegate is set at the very early state by AccessibilityChecker
(subclass of ChromeViewsDelegate) in ctor of InProcessBrowserTest. Introduced in v68.
To have a test, explicitly set views delegate by initializing here.
This is only effective in release mode because initializing views delegate twice isn't
allowed by dcheck."

In c76 the crash happens due to us initializing BraveViewsDelegateLinux in the test.
To work around it, override AccessibilityChecker to use BraveViewsDelegateLinux when
building browser tests on linux. This also allows to run the test in debug(non-official)
build.