Fix brave theme regression
1. Fix WebUI theme is changed to dark when tor/guest window is created.
When Tor window is created its BraveThemeService set system theme as a dark
To fix this system theme type is not updated when tor/guest profile is created.
this update is only done when normal profile is created.

2. Fix Tor/guest native ui is changed when brave theme in normal window is changed.
This is happened becase Tor/guest window uses NativeTheme.
To fix this, NativeThemeDarkAura is used for them.

3. Fix linux on X11 didn't use NativeThemeDarkAura for tor/guest profile properly.
ChromeBrowserMainExtraPartsViewsLinuxX11 is subclass of ChromeBrowserMainExtraPartsViewsLinux
and ChromeBrowserMainExtraPartsViewsLinux should be replaced with BraveBrowserMainExtraPartsViewsLinux.