Modified Transifex script to strip ph children.
In each .xtb file we create, strip all children from <ph> tags and
remove innter text if it start with $ (so that $1 text would be removed,
but something like '&lt;a href="$1"...' would not. This is needed so
that multiple replacement placeholders aren't generated by the resource
data pack loader.