SimplifyHttpsIndicator feature is now on by default.
We were overriding base::GetFieldTrialParamValueByFeature function call
in location_bar_model_impl.cc in chromium_src to force the experiment
outcome to
OmniboxFieldTrial::kSimplifyHttpsIndicatorParameterBothToLock.
The feature is now on by default and the behavior is the same as what we
were forcing. Removed the no longer needed override.

Chromium change:

https://chromium.googlesource.com/chromium/src/+/c26ef30d0961cbcdf2a226792922d73e6a1fc59f

commit c26ef30d0961cbcdf2a226792922d73e6a1fc59f
Author: Christopher Thompson <cthomp@chromium.org>
Date:   Fri Jun 7 01:55:45 2019 +0000

    Enable SimplifyHttpsIndicator feature by default

    This also cleans up the various old conditions, keeping the feature
    flag as a simple binary flag (enabled = lock-only). Also updates tests
    and removes the console warning and testing config.

    Bug: 803138