Do not load extensions in non-incognito guest profile
Two guest profiles are created. First, a regular on-the-record Profile is created, and then the off-the-record version of that Profile is created and returned to be used by a window. However, we were previously allowing extensions to be loaded in both these profiles, resulting in background scripts loading twice and receiving the same tab events.