Disable kLookalikeUrlNavigationSuggestionsUI
It used to be disabled by default, but enabled by experiments,
so we explicitly disabled it. When we got rid of experiments
we removed the explicit disablement as it was just disabled by
default. Now Chromium flipped it to be enabled by default, so
we need to explicitly disable it again.

Removed from being explicitly disabled by us in this commit:

commit 78c8700
Author: Pranjal Jumde <pranjal.jumde@gmail.com>
Date:   Tue May 21 00:04:17 2019 -0700

    Issue 4552: Restore features to correct state after disabling field trials

Chromium change:

https://chromium.googlesource.com/chromium/src/+/0a4ecf746ae7ffb2659ae2857a730ea17c68f754

commit 0a4ecf746ae7ffb2659ae2857a730ea17c68f754
Author: Mustafa Emre Acer <meacer@chromium.org>
Date:   Wed Jun 26 23:29:50 2019 +0000

    Enable Lookalike URLs interstitial by default

    This CL enables the interstitial by default for lookalike domains in
    user's navigation history (i.e. site engagement). Matching with top
    domains is controlled by the topsites parameter.