NTP and Sync WebUI do not use Toggle from Shields inside brave-ui
Shields will be removed from brave-ui and placed wholly inside brave-core, so the shields/Toggle component will not be available for use.
Ideally each web app should be self-contained apart from explicitly shared components and not use components from eachother. Since this is already the case, we change the import path to be correct. But Sync should change to use its own Toggle, or a shared Toggle.