Fix rewards extension action on start.
Make sure the rewards extension action only shows up when the rewards
service is ready (loaded ledger).

This is now the real fix to why the task to load ledger takes so long,
but it is good to have a mechanism in place to control the action based
on the service availability.