Hack wrapper to work around segfault (brave/brave-browser#4142)
The `brave` binary returns a non-zero error code and that causes
the wrapper to return a non-zero error code too. Since it's not
possible to override the error code with an `exec` call (`exec`
will terminate the shell immediately with a successful return
value), we need to keep the original shell around so that we can
hide the true return value of the `brave` binary.

This hack should be removed once we have fixed the crash.