Copy the notarized app to the output directory
The sign_all function in pipeline.py indicates that if package_dmg
is False it will copy the signed app to the paths.output directory
(which should be out/Release/signing in Brave,) but it doesn't
appear to do this. Here we add the step to copy it to this
location for use by the create_pkg and other downstream targets.