Fix launching Tor process when close and open Tor windows again.
Tor process is only killed on browser shutdown when it was based on guest
profile, and before https://github.com/brave/brave-core/pull/3597 is merged.
The guest profile and the dummy Tor regular profile are shutting down only
when closing the browser.
In this PR, the logic is fixed to shutdown Tor process whenever all Tor windows
are closed and properly relaunch it when creating a new Tor window.