Fix default search logic for Tor window and guest window
Fixes https://github.com/brave/brave-browser/issues/6476

Specifically, fixes the engine when a previously set engine is no longer available.

For example, in DE the engine used to be DDG (501), but was changed in
the prepopulated list to DDG_DE(516). The value in the preferences will
not be able to be used to get a valid engine (and cause a crash). This
change checks if the currently selected engine is available and if not
resets the value to the currently preset default.