Fix brave/brave-ios#1129: Match iOS 13's auto-play WebKit Policies (brave/brave-ios#1735)
Changes the default value to match the iOS 13 WebKit policies.

If a video has an audio channel, it shall NOT auto play! The default value as of iOS 13 is `.audio`. It's better to not to specify this value because if it changes again, our code will automatically have the applied effect.