Fixes clear cookies on exit toggle.
Fixes brave/brave-browser#6119

Account for CONTENT_SETTING_SESSION_ONLY as a possibility for the
default cookie setting in preferences - do not override it with
CONTENT_SETTING_ALLOW if it's already set.

Added browser test CookiePrefServiceTest.CookieControlType_Preference to
test Shields cookie setting change when preference setting changes
and in reverse.