Avatar button changes.
ShouldShowGenericIcon was removed and generic profile is now determined
by IsGenericProfile which sadly is scoped to the .cc file. Instead of
patching it overriding GetState which determines the type of profile. If
the profile is determined to be generic, then do our normal check and
change it to "normal".

Chromium change:

https://chromium.googlesource.com/chromium/src/+/a3b53120fca4a42f19afe4af29bc67a404581ec4

commit a3b53120fca4a42f19afe4af29bc67a404581ec4
Author: Jan Krcal <jkrcal@chromium.org>
Date:   Fri Sep 6 10:55:43 2019 +0000

    [Avatar button] Extract the UI state into one enum

    This CL refactors the avatar toolbar button so that the UI state is
    extracted into one state getter which results in consistent treatment
    across different functions displaying the UI and simplifies the code.

    Bug: 967317