Fix the issue, and add a unit test
Resolves #6545

The code in PhaseTwo::PrepareBatchCallback assumed that surveyor IDs were
unique, but they can be reused by different transactions. We were assigning
prepareBallot objects meant for one viewing ID to ballots for another viewing
ID. This caused the Anonize library to fail its signature checks.

To fix this, pass the viewing ID for the current transaction when calling
PhaseTwo::PrepareBatchCallback, and ensure that only ballots for that viewing
ID are updated with prepareBallot objects.