Fix #6607, and add a unit test.
Resolves #6607: Ballots assigned incorrect proof.

The code in PhaseTwo::ProofBallots assumed that surveyor IDs are unique, but
they may be reused between transactions. This could cause the proof for a
ballot with one viewing ID to be assigned to a ballot for a different viewing
ID if they shared the same surveyor ID.

To fix this, check that both the surveyor ID and viewing ID are a match before
assigning the proof.