Rewards button pre-opt-in badge dismisses after first button press
We use a new pref to store the "badge text" value. Whilst we could store a boolean value whose state would be "has dismissed notification" (or not), this method is more future proof for having multiple variations of badge text, especially if or when this button becomes 'native' all the time (i.e. not an extension icon where the badge is controlled via JS)