FlagsState refactored.
Linking browser tests on Linux was erroring out with duplicate symbols.
We were including chrome/browser/about_flags.cc and
components/flags_ui/flags_state.cc into the chromium_src override of
about_flags.cc. This caused duplication of symbols because these files
are in different layers and are built into different libraries.

The refactor is such that about_flags just calls into flags_state
singleton, so can only modify about_flags with an override. But, since
the original calls a method with the same name in flags_state, added an
alias for it in there.

Chromium change:

https://chromium.googlesource.com/chromium/src/+/3cd752835702327d76cc10fb04f4fe0ae7c8c702

commit 3cd752835702327d76cc10fb04f4fe0ae7c8c702
Author: Elly Fong-Jones <ellyjones@chromium.org>
Date:   Tue Nov 12 20:26:50 2019 +0000

    flags: refactor FlagsState interface

    This change:
    1) has FlagsState take (and internally store) its set of FeatureEntry as
       a base::span<FeatureEntry> rather than a (FeatureEntry*, size_t) pair,
       and rewrites all iterations over the set of FeatureEntry to use
       new-style for loops.
    2) introduces FlagsState::Delegate for users of the component to
       customize the behavior of FlagsState, and replaces existing uses of
       the exclude predicate with uses of this delegate.

    This is a preparatory refactor for the work to show flag expiration
    milestones within the flags UI. As that work proceeds
    FlagsState::Delegate will gain new methods.

    Bug: 1020637