Fixes rewards-image handler post C80 bump.
The handler now gets a full url instead of just path. In the past the
path was the actual url we wanted retrieved, but now the url includes
the chrome://rewards-image prefix, so we need to extract path from the
url. URLDataSource::URLToRequestPath is used to do that.

Also updated NTP Sponsored Image handler to use
URLDataSource::URLToRequestPath instead of extracting path manually.

Fixes brave/brave-browser#8074