More Mac signing script changes.
Moved call to create Brave config into config_factory.py.
Added a static method to our config creation to prevent an exception
when Chrome's config loading is attempted.

Maybe need to add code to throw an exception when provisioning profile
is not found from the env when not a development config, but development
var isn't passed into the config_factory which means we'd need to add
another patch.

Chromium change:

https://chromium.googlesource.com/chromium/src/+/f21027ab4d74ea035d98bbf6ba77e8fa148f08b5

commit f21027ab4d74ea035d98bbf6ba77e8fa148f08b5
Author: Joshua Pawlicki <waffles@chromium.org>
Date:   Mon Mar 23 17:45:15 2020 +0000

    Create an updater-specific signing pipeline.

    Bug: 926234