Fixed crash when user tries to install ipfs extension from setting
When extension install is requested, install dialog is launched.
So far, we requsted un-parented install dialog.
On linux, custom frame is not allowed for un-parented dialog.
Only platform-native frame is used for it. The problem was
install dialog tries to set title to custom frame. So, seg fault
happened only on linux. And Windows with classic theme will have
same issue. On macOS and Windows with aero, crash doesn't happen.
But, un-parented dialog has title area that looks ugly.