Fixes site settings Autoplay value selector.
Fixes brave/brave-browser#10049

Content settings types list in constants.js is meant to be kept in sync
with chrome/browser/ui/webui/site_settings_helper.cc. We do not patch
the enum in site_settings_helper.cc, but instead override functions that
manipulate it.

On js side we can try to override the behavior, but that may introduce
more patches than just patching the constants.