test: Simulate Sentinel value caching
CreateSentinelIfNeeded() is called in
chrome/browser/chrome_browser_main.cc, and that calls
GetFirstRunSentinelCreationTime() and caches the result of that
operation. This means when we call the function again in
BraveStatsUpdaterParams, the call will never block. Change the tests to
have this behavior, because right now tests are timing out.

Resolves https://github.com/brave/brave-browser/issues/10130