[Android] site_settings moved to components
Chromium change:

https://chromium.googlesource.com/chromium/src/+/e012e41e7a0648468a018f54f963f34efb0eaf9c

This CL moves the majority of //chrome/browser/site_settings to
//components/browser_ui/site_settings, leaving only Chrome-specific
classes behind. This is part of an effort to componentize site_settings
so it can be reused by WebLayer.

This CL leaves most of the tests in //chrome because they rely on the
SettingsActivity and specific chrome implementation of the site
settings clients. I felt that componentizing these tests would result
in removing enough of the chrome based scaffolding around the UI that
it would reduce the effectiveness of the tests.

Bug: 1058595