[Browser Tests] Modified RewardsFlagBrowserTest tests.
One of the tests - RewardsFlagBrowserTest.HandleFlagsWrongInput - was
failing on both Windows and Linux in Release builds, but succeeded in
Debug. I believe this is due to the test using RunLoop::RunUntilIdle to
wait for async callbacks. As the documentation for RunUntilIdle notes:

// WARNING #2: This may return too early! For example, if used to run until an
//             incoming event has occurred but that event depends on a task in
//             a different queue -- e.g. another TaskRunner or a system event.
// Per the warnings above, this tends to lead to flaky tests;

In the debugger it was clear that the callbacks were getting called
after the RunUntilIndle returned and the test was exiting prematurely.

Replaced all waits via RunUntilIdle with RunLoop's Run/Quit methods.

Fixes brave/brave-browser#10392